Organize package lib
- Move shim stuff into lib/shim/
  - Move lib/web3-shim to lib/shim/index
  - Move overloads lib/*-overloads to lib/shim/overloads/*
- Move adapter stuff into lib/adapter/
  - Move lib/interface-adapter.ts to lib/adapter/index.ts
  - Move lib/interface-adapter/types/index.d to lib/adapter/types
  - Rename lib/interface-adapter to lib/adapter (better autocomplete)
  - Move lib/web3-interface-adapter to lib/adapter/web3
