+++ packages/interface-adapter/lib/adapter/types.ts
