+++ packages/interface-adapter/lib/adapter/web3/index.ts
@@ -1,5 +1,5 @@
-import { Web3Shim, Web3ShimOptions } from "./web3-shim";
-import { EvmBlockType } from "./interface-adapter/types";
+import { Web3Shim, Web3ShimOptions } from "../../shim";
+import { EvmBlockType } from "../types";
 import { Provider } from "@truffle/provider";
 
 export interface Web3InterfaceAdapterOptions extends Web3ShimOptions {}
