+++ packages/interface-adapter/lib/shim/overloads/ethereum.ts
@@ -1,10 +1,10 @@
 import BN from "bn.js";
-import { Web3Shim } from "./web3-shim";
+import { Web3Shim } from "..";
 import {
   Transaction as EvmTransaction,
   TransactionReceipt as EvmTransactionReceipt
 } from "web3-core";
-import { EvmBlockType } from "./interface-adapter/types";
+import { EvmBlockType } from "../../adapter/types";
 
 export const EthereumDefinition = {
   async initNetworkType(web3: Web3Shim) {
