+++ packages/interface-adapter/lib/shim/overloads/web3js.ts
@@ -1,4 +1,4 @@
-import { Web3Shim } from "./web3-shim";
+import { Web3Shim } from "..";
 
 // We simply return plain ol' Web3.js
 export const Web3JsDefinition = {
